Landing page
The first landing page completed to the end.
The basis of the design is taken from the free layout.
I developed the design of the mobile version and adaptive design.
I tried to use BEM, and only at the end of the work it became clear how to do it even better.
Worked on this for 4 days. Gallery not implemented. I would like to write vue.js later (when learn it).
Do not wait to realize the accumulated knowledge in the next project.